Dedupe overflow-test setup into overflow_test_utils.py
skip_if_insufficient_gpu_memory and wide_noncontiguous_slices were only
shared within test_ssd.py; test_layer_norm.py and test_layernorm_gated.py
independently duplicated the backward-kernel row_start overflow-margin
calculation. Factored the shared pieces into a plain (not underscore-
prefixed -- no such convention exists in this repo) helper module.
